feat(scripts): Add write-userconfig
