+++ packages/uhk-usb/src/models/load-configurations-result.ts
@@ -1,0 +1,4 @@
+export interface LoadConfigurationsResult {
+    userConfiguration: string;
+    hardwareConfiguration: string;
+}
